add resources locales and add init.js
