+++ index.html
@@ -39,8 +39,8 @@
           <div class="input-button-container">
         <textarea placeholder="Send a message" cols="110" rows="0"> </textarea>
           <div id="answer" class="hidden">  <!--   Код для отображения ответа  - работает ли>? если что удалить и в html и css --></div>
-        <button class="btn-submit" >➤</button>
-        <button class="btn-trash" >🗑️</button>
+        <button class="btn-submit" disabled>➤</button>
+        <button class="btn-trash" disabled>🗑️</button>
         </div>
       </form>
       </div>
