+++ src/locales/en.js
@@ -1,0 +1,8 @@
+export default {
+    translation: {
+        themeDark: 'Dark mode',
+        themeLight: 'Light mode',
+        reset: 'Regenerate response',
+        addChat:'New Chat',
+      },
+}
