+++ src/locales/index.js
@@ -1,0 +1,4 @@
+import en from './en.js';
+import ru from './ru.js';
+
+export default {en, ru};
