+++ src/locales/ru.js
@@ -1,0 +1,8 @@
+export default {
+    translation: {
+        themeDark: 'Темный режим',
+        themeLight: 'Светлый режим',
+        reset: 'Перегенерировать ответ',
+        addChat:'Новый чат',
+    },
+}
